initial commit
Thought is that every monoid should be isomorphic to a monoid
that has a one part applied. Something like:

Every Monoid m isomorphic to a Monoid (m -> m)

Example (Point, +) isomorphic to (Vector, <>)
If we think of a Vector as a special type of function that has
the signature (Point -> Point).  In this case a Vector is a partially
applied Point + operator.  What is more is that the compose can be
done differently than regular function composition.  Composing two
Vector functions should yield a single Vector, and not the normal
successive application of two functions.

(<> is supposed to look like function composition operator)